Move all messaging patterns to device specification (#194)
To further clarify and separate the specification of device operation
from the binary protocol, we move all messaging patterns to Device.md.

This greatly reduces the number of cross-references and duplicate
terminology across both documents.

* Remove ambiguous versioning string
* Move error handling to messaging patterns section
* Clarify rejection of register writes
* Account for possibility to flag events as errors

In the near-future we may want to use this as an explicit exception
throwing mechanism used by the Parse operator.

* Clarify error handling recommendations
* Improvements for register and serial interface
* Improve wording for describing connection checks
* Improve wording to clarify response contents

Specifically, we want to clarify the payload should send the register
contents after the request is processed.

* Fix broken header link

---------

Co-authored-by: bruno-f-cruz <7049351+bruno-f-cruz@users.noreply.github.com>